// demo-mcp-21-code-generator-node/input/reference-repos/ECOMMERCE_V9/MSAbCaAlarPricesETL/src/common/utils/enums/tracing.enum.ts
/**
 * Enumeraciones usadas internamente para la trazabilidad y descripción de tareas en los logs del microservicio
 * @author Fredy Santiago Martinez
 */

export enum EStatusTracingGeneral {
  FAILED = 'FAILED',
  ERROR = 'ERROR',
  STATUS_SUCCESS = 'SUCCESS',
  STATUS_IN_PROCESS= 'STATUS_IN_PROCESS',
  STATUS_FAILED = 'FAILED',
  LEGACY_SUCCESS = 'LEGACY_SUCCESS',
  LEGACY_ERROR = 'LEGACY_ERROR',
  LEGACY_WARN = 'LEGACY_WARN',
  BD_SUCCESS = 'BD_SUCCESS',
  BD_ERROR = 'BD_ERROR',
  BD_WARN = 'BD_WARN',
}

export enum ETaskTracingGeneral {
  INVALID_CHANNEL = 'INVALID CHANNEL',
  DOWNLOAD_DATA = 'DESCARGA DE DATOS',
  SAVE_DATA_MOVIL = 'GUARDADO DE DATOS',
  DELETE_DATA_MOVIL = 'ELIMINACION DE DATOS',
  TASK_GENERAL = 'TEST',
  EXEC_CRON = 'EJECUCION CRON',
  EXEC_MANUAL = 'EJECUCION MANUAL',
  SAVE_DATA = 'GUARDADO DE DATOS',
  DELETE_DATA = 'ELIMINACION DE DATOS',
  FILE_DOWNLOAD = 'DESCARGANDO ARCHIVO',
  INICIO_REQUEST = 'INICIO_REQUEST',
  FINAL_REQUEST = 'FINAL_REQUEST',
}

export enum EDescriptionTracingGeneral {
  START_JOB_PROCESS = 'INICIA PROCESO DE JOB PRICESETL',
  START_MANUAL_PROCESS = 'INICIA PROCESO MANUAL PRICESETL',
  FINISH_JOB_PRICES = 'FINALIZA PROCESO PRICES DISPONOBILITY',
  START_CALL_ORCHESTRATOR = 'INICIA LLAMADO A ORQUESTADOR',
}